backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
openai
google-generativeai
Pillow
